build: Include pugixml in installs
When we build using FetchContent and install the result, we have to provide the
dependencies as well, or we will fail in a subsequent step when we try to find
pugixml via cmake find_package at the enduser systems.

Note: there is always the case where we can use system supplied libraries via
FETCHCONTENT_FULLY_DISCONNECTED = ON
with this we only have to install our libraries.

Signed-off-by: Tobias Klausmann <tobias.klausmann@lenze.com>