MinMaxString: Introduce new common base class
This will be used in coming commits

The new V3 Metamodel imposes restrictions on string min and max length. We have
to follow those to be complient, thus we need to constrain our string to the
imposed sizes. This class is introduced to fulfill the need for constrained
strings.

Strings constrained by the new standard are (min/max characters):
Identifier: min: 1, max: 2000
LabelType: min: 1, max 64

Signed-off-by: Tobias Klausmann<tobias.klausmann@lenze.com>